requirements.txt
praw==7.8.1
openai==1.95.1
python-dotenv==1.0.1